requirements.txt
numpy==1.23.5
scipy==1.15.3
pandas==2.3.3
scikit-learn==1.7.2
salesforce-merlion==1.1.0